change specs to run against a test company; added time filter specs
The specs now require a test.qbw company file to be placed in the same
directory as the specs. This helps to prevent any accidental damage to a
real company file. Also, if the specs are run on a non-Windows machine,
a warning will be displayed, and any of the specs that require access to
QuickBooks will be skipped.

Specs have been added for the created_before and created_after filters.